refactor: move filters to header
